Dropped outdated arcfour cipher (#12)
OpenSSL 7.6 removed support for arcfour. rsync which uses ssh is trying
to establish connection with use of it. For newer installations this process
is ending with an error to establish connection.
Replaced arcfour with aes128-gcm@openssh.com to allow it to work.
However, user can overwrite this by providing --cloud-rsync-cipher argument